made bosses do shit

+++ SmallGameProject/SmallGameProject/Abilities.h
@@ -12,7 +12,7 @@ protected:
 public:
 	ArcFire();
 	virtual~ArcFire();
-	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, int triggerDelay);
+	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, float triggerDelay);
 	virtual float activate(Enemy* enemy, EntitySubject* entitySubject, DirectX::XMFLOAT3 playerPos);
 	virtual void update(Enemy* enemy, EntitySubject* entitySubject, float deltaTime);
 
@@ -30,7 +30,7 @@ protected:
 public:
 	SplitFire();
 	virtual~SplitFire();
-	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, int triggerDelay, int projectilesOnSplit, float splitArc);
+	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, float triggerDelay, int projectilesOnSplit, float splitArc);
 	virtual float activate(Enemy* enemy, EntitySubject* entitySubject, DirectX::XMFLOAT3 playerPos);
 	virtual void update(Enemy* enemy, EntitySubject* entitySubject, float deltaTime);
 
@@ -45,7 +45,7 @@ protected:
 public:
 	ReverseFire();
 	virtual~ReverseFire();
-	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, int triggerDelay);
+	virtual void Initialize(float arc, int nrOfProjectiles, float cooldown, int attackDelay, int maxCharges, float triggerDelay);
 	virtual float activate(Enemy* enemy, EntitySubject* entitySubject, DirectX::XMFLOAT3 playerPos);
 	virtual void update(Enemy* enemy, EntitySubject* entitySubject, float deltaTime);
